moving Footer to common components folder
