+++ src/app/(unauthenticated)/layout.tsx
@@ -1,6 +1,6 @@
 import Box from '@mui/material/Box';
 import { UnauthenticatedNavbar } from '@/app/components/common/Navbars';
-import { Footer } from '@/app/components/Footer';
+import { Footer } from '@/app/components/common/Footer';
 import { AlertBarProvider, AlertBar } from '@/app/components/common/AlertBar';
 
 export default function UnauthenticatedLayout({ children }: { children: React.ReactNode }) {
